MenuPage get data from db and links to category page
