+++ src/App.jsx
@@ -1,14 +1,10 @@
 // packages
-import { useEffect, useState } from "react";
 import {
   BrowserRouter as Router,
   Routes,
   Route,
   Navigate,
 } from "react-router-dom";
-
-// scripts
-import { getCollection } from "./scripts/fireStoreDB";
 
 // components
 import HomePage from "./pages/HomePage";
@@ -18,22 +14,6 @@ import MenuPage from "./pages/MenuPage";
 import AdminPage from "./pages/AdminPage";
 
 export default function App() {
-  const [menu, setMenu] = useState([]);
-  const [status, setStatus] = useState(0);
-
-  useEffect(() => {
-    async function loadData(path) {
-      const data = await getCollection(path);
-
-      setMenu(data);
-      setStatus(1);
-    }
-    loadData("categories");
-  }, []);
-
-  if (status === 0) return <p>Loading...</p>;
-  if (status === 2) return <p>Error!</p>;
-
   return (
     <Router>
       <Routes>
@@ -41,11 +21,8 @@ export default function App() {
         <Route path="/*" element={<Navigate replace to="/" />} />
         <Route path="admin" element={<AdminPage />} />
         <Route path="menu" element={<MenuPage />} />
-        <Route path="menu/:course/:category" element={<CategoryPage />} />
-        <Route
-          path="menu/:course/:category/:productId"
-          element={<ProductPage />}
-        />
+        <Route path="menu/:category" element={<CategoryPage />} />
+        <Route path="menu/:category/:productId" element={<ProductPage />} />
       </Routes>
     </Router>
   );
